Update wipe.sh look & feel

+++ ali/utils/wipe.sh
@@ -7,22 +7,31 @@
 # +--------------------------------------------+
 
 # Root privileges required
-[[ "${UID}" -ne 0 ]] && exit 0
+[[ "${UID}" -ne 0 ]] && exit 0 || clear
+
+# Colors
+OFF='\e[0m'
+CYAN='\e[0;36m'
+BLUE='\e[1;34m'
 
 # Hard Disk Drive label
 HARDDISK='/dev/sda'
 
 # Zero-fill
+echo -e "${BLUE}:: Zero-fill device: ${CYAN}${HARDDISK}\n${OFF}"
 dd if=/dev/zero of=${HARDDISK} bs=1M status=progress
 
 # Diagnostic
+echo -e "${BLUE}\n:: Smart diagnostic: ${CYAN}${HARDDISK}\n${OFF}"
 smartctl --test=short ${HARDDISK} && while [[ $(smartctl --all ${HARDDISK}) =~ 'progress' ]]; do sleep 10; done
 
 # Statistics
+echo -e "${BLUE}:: Smart statistics: ${CYAN}${HARDDISK}\n${OFF}"
 smartctl --health ${HARDDISK}
 smartctl --log=error ${HARDDISK}
 smartctl --log=selftest ${HARDDISK}
 
-# SMART status
+# Status
+echo -e "${BLUE}:: Smart support: ${CYAN}${HARDDISK}\n${OFF}"
 smartctl --info ${HARDDISK} | grep 'SMART support'
